requirements.txt
python-can
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
websockets
pydantic
tensorflow
